fix(released): upgrade download method sftp.

+++ sample/BERT/scripts/download.sh
@@ -1,31 +1,34 @@
 #!/bin/bash
-pip3 install dfn
-# sudo apt install unzip
-
+res=$(dpkg -l|grep unzip)
+if [ $? != 0 ];
+then
+    echo "Please install unzip on your system!"
+    exit
+fi
+pip3 install dfss -i https://pypi.tuna.tsinghua.edu.cn/simple --upgrade
 scripts_dir=$(dirname $(readlink -f "$0"))
-# echo $scripts_dir
 
 pushd $scripts_dir
 # datasets
 if [ ! -d "../datasets" ]; 
 then
-    python3 -m dfn --url http://disk-sophgo-vip.quickconnect.cn/sharing/a0geo4gVx
-
+    python3 -m dfss --url=open@sophgo.com:sophon-demo/BERT/datasets_0918/datasets.zip
     unzip datasets.zip -d ../
     rm datasets.zip
+
     echo "datasets download!"
 else
-    echo "datasets exist!"
+    echo "Datasets folder exist! Remove it if you need to update."
 fi
 
 # models
 if [ ! -d "../models" ]; 
 then
-    python3 -m dfn --url http://disk-sophgo-vip.quickconnect.cn/sharing/u8EzJBIA7
+    python3 -m dfss --url=open@sophgo.com:sophon-demo/BERT/models_0918/models.zip
     unzip models.zip -d ../
     rm models.zip
     echo "models download!"
 else
-    echo "models exist!"
+    echo "Models folder exist! Remove it if you need to update."
 fi
 popd
